% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decapoda.R
\docType{data}
\name{Decapod.dat}
\alias{Decapod.dat}
\title{Decapod crustaceans richness dataset}
\format{
The dataset comes with variables:
\describe{
     \item{Richness}{Summed abundance of 12 families expressed as number of individuals per m^2}
     \item{Year}{Integer denoting year}
     \item{Location}{Integer denoting location}
     \item{T1m}{Numeric, temperature at 1 m depth in °C}
     \item{T45.35m}{Temperature at 45-35 m depth in °C}
     \item{S1m}{Salinity at 1 m depth}
     \item{S45.35m}{Salinity at 45-35 m depth}
     \item{Ch0.10m}{Chlorophyll a at 10 m depth}
}
}
\usage{
data("decapoda")
}
\description{
Richness dataset of decapod crustaceans with additional explanatory variables
including year, location, salinity, temperature and chlorophyll a concentrations.
}
\details{
Decapods are an important fishery and of ecological importance. One can find a wide
diversity in coastal ecosystems. Many produce pelagic larvae that eventually settle
into benthos. This dataset comes from a PhD project investigating ecological role
and recruitment of decapod larvae. For the purpose of the book, only 12 decapod
crustacean families are studied over two years.
}
\references{
See FRS (Fisheries Research services) long-term weekly monitoring programme 
referenced in the book.

Zuur et al. 2009. Mixed effects models and extensions in ecology with R. Springer. ISBN: 978-0-387-87457-9.
}
