% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nereis.R
\docType{data}
\name{Nereis}
\alias{Nereis}
\title{Nereis data}
\usage{
data("nereis")
}
\description{
Data of deposit-feeding \emph{Nereis (Hediste) diversicolor} from Ieno et al. (2006).
}
\details{
Data of deposit-feeding \emph{Nereis (Hediste) diversicolor} comes from Ieno et al. (2006).
 Data was collected to model nutrient concentration.
}
\references{
Ieno et al. 2006. Distinguishing between the effects of infaunal species richness, identity and density in the marine benthos. Marine ecology progress series, 311: 263-271.
}
