% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seals.R
\docType{data}
\name{Seals}
\alias{Seals}
\title{Seals data}
\usage{
Seals
}
\description{
Common seal (\emph{Phoca vitulina} L.) data comes from Cronin et al. 2007.
}
\details{
Common seal (\emph{Phoca vitulina} L.) data comes from Cronin et al. 2007, where between April 2003 and November 2005, regular
standardised haul-out count survey of Bantry Bay and Kenmare River was carried out by boat. Counts of seals at each haul-out
site was carried out independently and simultaneously by two observers, initially from 200 m and progressively closer. Read
more about the data in Zuur et al. 2007 in chapter 23 (page 503) or in above mentioned Cronin et al. 2007 paper.
}
\references{
Cronin et al. 2007. An assessment of population size and distribution of harbor seals in the Republic ofIreland during the moult season in August 2003. Journal of zoology, 273: 131-139

Zuur et al. 2009. Mixed effects models and extensions in ecology with R. Springer. ISBN: 978-0-387-87457-9.
}
