% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solea.R
\docType{data}
\name{Solea}
\alias{Solea}
\title{Solea solea dataset}
\source{
Unable to find the source, see the book for details?
}
\usage{
data("Solea")
}
\description{
Absence and presence of Solea solea flatfish in Tagus estuary.
}
\details{
Flatfish (Solea solea) presence and absence data

The data is abundance of \emph{Solea solea} in the Tagus estuary (Portugal),
along with some of the abiotic factors measured at the location.
The \emph{Solea solea} abundance data was collected monthly in four areas
in 1995 and 1996 using 4 m beam trawl with 10 mm mesh and one tickler chain.

The data consists of 65 observations and 13 variables.

\describe{
 \item{Sample}{Unique ID of the observation.}
\item{season}{Season of the sampling, either spring (1) or summer (2).}
\item{month}{Month of the sampling.}
\item{Area}{Four areas where the sampling took place.}
\item{depth}{Depth of sampling in meters.}
\item{temperature}{Temperature during sampling in °C.}
\item{salinity}{Salinity at the sampling site in ppt.}
\item{transparency}{Water transparency in cm.}
\item{gravel}{Percentage of gravel in the sediment.}
\item{large_sand}{Percentage of large sand in the sediment.}
\item{med_fine_sand}{Percentage of medium and fine sand in the sediment.}
\item{mud}{Percentage of mud in the sediment.}
\item{Solea_solea}{Presence or absence of Solea solea in the sample.}
}
}
\references{
Zuur, A., Ieno, E.N. and Smith, G.M., 2007. Analyzing ecological data. Springer Science & Business Media.
}
