% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antarctic_birds.R
\docType{data}
\name{AntarcticBirds}
\alias{AntarcticBirds}
\title{Antarctic birds and climate change}
\usage{
data("antarctic_birds")
}
\description{
Antartic birds (\emph{Aptenodytes forsteri}, \emph{Pygoscelis adeliae}, \emph{Daption capense}) and
their phenological data along with some environmental data.
}
\details{
Dataset contains information on three bird species frequenting Antartic, the Emperor penguin (\emph{Aptenodytes forsteri}), 
Adelie penguin (\emph{Pygoscelis adeliae}) and Cape petrel (\emph{Daption capense}). Variables in the dataset include
below variables. Arrival and laying times are days in the year.
\describe{ 
    \item{ArrivalAP}{Arrival time of Adelie penguin to the breeding area}
    \item{LayingAP}{Laying time of Adelie penguin}
    \item{ArrivalCP}{Arrival time of Cape petrel to the breeding ground}
    \item{LayingCP}{Laying time of Cape petrel}
    \item{ArrivalEP}{Arrival time of Emperor penguin}
    \item{LayingEP}{Laying time of Emperor penguin}
    \item{MSA}{Methanesulfonic acid, a proxy for sea ice extent since satellite images are available only from 1970
    onwards}
    \item{SOI}{Southern oscillation index, also known as El Niño souther oscillation conditions. High positve values
     indicate La Niña and low negative values indicate El Niño, which have been shown to have impact on demographic
     rates and food resources of many animals, including seabirds.}
    }
}
\references{
Barbraud C. and Weimerkirch H. 2006. Antarctic birds breed later in response to climate change.
Proceedings of the National Academy of Sciences of the USA 103: 6048-6051.

Zuur et al. 2009. Mixed effects models and extensions in ecology with R. Springer. ISBN: 978-0-387-87457-9.
}
