% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RIKZ.R
\docType{data}
\name{RIKZ}
\alias{RIKZ}
\title{Marine benthic data}
\usage{
data("RIKZ")
}
\description{
Marine benthic data comes from Zuur et al. (2007).
}
\details{
Marine benthic data from nine inter-tidal areas along the Dutch coast. The 
data were collected by the Dutch institute RIKZ in the summer of 2002. In 
each inter-tidal area (denoted by ‘beach’), five samples were taken, and the 
macro-fauna and abiotic variables were measured.

\itemize{ 
    \item Sample: sample number
    \item Richness: species richness
    \item Exposure: index composed of the surf zone, slope, grain size, and depth of anaerobic layer
    \item NAP: height of sampling station compared to mean tidal level
    \item Beach: beach identifier
    }
}
\references{
Zuur, A., Ieno, E.N. and Smith, G.M., 2007. Analyzing ecological data. Springer Science & Business Media.
}
