% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildboar_tb.R
\docType{data}
\name{WildBoarTb}
\alias{WildBoarTb}
\title{Wild boar tuberculosis-like lesions dataset}
\usage{
data("wildboar_tb")
}
\description{
Presence or absence of tuberculosis in wild boar (\emph{Sus scrofa}) from south central Spain
}
\details{
Vicente et al. (2006) analyzed the distribution on tuberculosis-like lesions in wild boar (\emph{Sus scrofa}) for
potential importance of persistence of tuberculosis in south central Spain. Data came from harvested animals.
Tuberculosis (Tb) is a zoonosis that can jump from non-human to human animal, which makes the disease interesting to
study for its economical and public health impacts.

The dataset comes with variables:
\itemize{
    \item Tb: An integer denoting presence (1) or absence (0) of tuberculosis in an animal
    \item sex: Sex of the animal harvested
    \item age_class: Age class of the animal based on tooth eruption patterns. Boars less than 7 months old were
    classified as piglets, boars between 7 and 12 months were classified as weaners, those between 12 and 24 months
    as juveniles, and those more than 2 years old as adults.
    \item length_ct: Head-body length of the animal in cm, where CT is Spanish abbreviation for \emph{cabeza-tronco}.
}
}
\references{
Vicente J, Höfle U, Garrido JM, Fernández-de-Mera IG, Juste R, Barralb M, Gortazar C. 2006. Wild boar
and red deer display high prevalences of tuberculosis-like lesions in Spain. Veterinary research 37: 107-119.

Zuur et al. 2009. Mixed effects models and extensions in ecology with R. Springer. ISBN: 978-0-387-87457-9.
}
